add auth check for delete

+++ app/Http/Controllers/TaskStatusController.php
@@ -101,6 +101,10 @@ class TaskStatusController extends Controller
      */
     public function destroy(TaskStatus $taskStatus)
     {
+        if (!Auth::check()) {
+            abort(419);
+        }
+
         $taskStatus->delete();
 
         flash(__('flash.task_statuses.delete.success'))->success();
